scheduling: merge simple and condition schedulers
